Changed return type of delete meta node
Even if delete meta node do not find the id it returns success.
http://dcgshare.lm.intel.com/reviews/r/482/

+++ dnvme_ds.h
@@ -100,7 +100,7 @@ struct metrics_sq {
 struct metrics_meta_data {
     struct list_head meta_trk_list;
     struct dma_pool *meta_dmapool_ptr;
-    u32    meta_buf_size;
+    u16    meta_buf_size;
 };
 
 /*
